// FrontBookRentals/src/app/api-client/model/loginResponseDtoGenericResponse.ts
/**
 * Book Rentals API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LoginResponseDto } from './loginResponseDto';


export interface LoginResponseDtoGenericResponse { 
    success: boolean;
    message: string;
    data?: LoginResponseDto;
}